+++ packages/backend/src/health.ts
@@ -1,6 +1,9 @@
 import {
     DbtProjectConfig,
     HealthState,
+    LightdashInstallType,
+    LightdashMode,
+    ProjectType,
     sensitiveDbtCredentialsFieldNames,
 } from 'common';
 import fetch from 'node-fetch';
@@ -37,6 +40,10 @@ export const getHealthState = async (
 
     const needsProject = !(await projectService.hasProject());
 
+    const localDbtEnabled =
+        process.env.LIGHTDASH_INSTALL_TYPE !== LightdashInstallType.HEROKU &&
+        lightdashConfig.mode !== LightdashMode.CLOUD_BETA;
+
     const defaultProject =
         needsProject && lightdashConfig.projects[0]
             ? (Object.fromEntries(
@@ -55,7 +62,11 @@ export const getHealthState = async (
         version: VERSION,
         needsSetup: !(await hasUsers(database)),
         needsProject,
-        defaultProject,
+        localDbtEnabled,
+        defaultProject:
+            !localDbtEnabled && defaultProject?.type === ProjectType.DBT
+                ? undefined
+                : defaultProject,
         isAuthenticated,
         latest: { version: latestVersion },
         rudder: lightdashConfig.rudder,